# generated/swagger.json
{
	"components": {
...
		"schemas": {
...
			"WeeklySummary": {
				"properties": {
					"workoutCompletionRate": {
						"type": "number",
						"format": "double"
					},
					"exerciseCompletionRate": {
						"type": "number",
						"format": "double"
					},
					"streak": {
						"type": "number",
						"format": "double"
					},
					"totalWorkoutsThisWeek": {
						"type": "number",
						"format": "double"
					},
					"completedWorkoutsThisWeek": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"workoutCompletionRate",
					"exerciseCompletionRate",
					"streak",
					"totalWorkoutsThisWeek",
					"completedWorkoutsThisWeek"
				],
				"type": "object",
				"additionalProperties": false
			},
			"WorkoutConsistency": {
				"properties": {
					"week": {
						"type": "string"
					},
					"weekLabel": {
						"type": "string"
					},
					"totalWorkouts": {
						"type": "number",
						"format": "double"
					},
					"completedWorkouts": {
						"type": "number",
						"format": "double"
					},
					"completionRate": {
						"type": "number",
						"format": "double"
					},
					"isInProgress": {
						"type": "boolean"
					},
					"status": {
						"type": "string",
						"enum": [
							"completed",
							"in-progress",
							"upcoming"
						]
					}
				},
				"required": [
					"week",
					"weekLabel",
					"totalWorkouts",
					"completedWorkouts",
					"completionRate",
					"isInProgress",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"WeightAccuracyMetrics": {
				"properties": {
					"accuracyRate": {
						"type": "number",
						"format": "double"
					},
					"totalSets": {
						"type": "number",
						"format": "double"
					},
					"exactMatches": {
						"type": "number",
						"format": "double"
					},
					"higherWeight": {
						"type": "number",
						"format": "double"
					},
					"lowerWeight": {
						"type": "number",
						"format": "double"
					},
					"avgWeightDifference": {
						"type": "number",
						"format": "double"
					},
					"chartData": {
						"items": {
							"properties": {
								"count": {
									"type": "number",
									"format": "double"
								},
								"color": {
									"type": "string"
								},
								"value": {
									"type": "number",
									"format": "double"
								},
								"label": {
									"type": "string"
								}
							},
							"required": [
								"color",
								"value",
								"label"
							],
							"type": "object"
						},
						"type": "array"
					},
					"hasPlannedWeights": {
						"type": "boolean"
					},
					"hasExerciseData": {
						"type": "boolean"
					}
				},
				"required": [
					"accuracyRate",
					"totalSets",
					"exactMatches",
					"higherWeight",
					"lowerWeight",
					"avgWeightDifference",
					"chartData"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GoalProgress": {
				"properties": {
					"goal": {
						"type": "string"
					},
					"progressScore": {
						"type": "number",
						"format": "double"
					},
					"totalSets": {
						"type": "number",
						"format": "double"
					},
					"totalReps": {
						"type": "number",
						"format": "double"
					},
					"totalWeight": {
						"type": "number",
						"format": "double"
					},
					"completedWorkouts": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"goal",
					"progressScore",
					"totalSets",
					"totalReps",
					"totalWeight",
					"completedWorkouts"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TotalVolumeMetrics": {
				"properties": {
					"date": {
						"type": "string"
					},
					"totalVolume": {
						"type": "number",
						"format": "double"
					},
					"exerciseCount": {
						"type": "number",
						"format": "double"
					},
					"label": {
						"type": "string"
					}
				},
				"required": [
					"date",
					"totalVolume",
					"exerciseCount",
					"label"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"WeeklySummaryResponse": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"error": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/WeeklySummary"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"WorkoutConsistencyResponse": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"error": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/WorkoutConsistency"
						},
						"type": "array"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"WeightAccuracyResponse": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"error": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/WeightAccuracyMetrics"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GoalProgressResponse": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"error": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/GoalProgress"
						},
						"type": "array"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TotalVolumeMetricsResponse": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"error": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/TotalVolumeMetrics"
						},
						"type": "array"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "mastersfit-server",
		"version": "1.0.0",
		"description": "RESTful API server for Masters Fit fitness tracking application",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/dashboard/{userId}/weekly-summary": {
			"get": {
				"operationId": "GetWeeklySummary",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WeeklySummaryResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WeeklySummaryResponse"
								}
							}
						}
					}
				},
				"description": "Get weekly summary for a user",
				"tags": [
					"Dashboard"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/dashboard/{userId}/workout-consistency": {
			"get": {
				"operationId": "GetWorkoutConsistency",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WorkoutConsistencyResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WorkoutConsistencyResponse"
								}
							}
						}
					}
				},
				"description": "Get workout consistency data for a user",
				"tags": [
					"Dashboard"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "startDate",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "endDate",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "timeRange",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"1w",
								"1m",
								"3m",
								"6m",
								"1y"
							]
						}
					}
				]
			}
		},
...
		"/dashboard/{userId}/weight-accuracy": {
			"get": {
				"operationId": "GetWeightAccuracy",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WeightAccuracyResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WeightAccuracyResponse"
								}
							}
						}
					}
				},
				"description": "Get weight accuracy metrics for a user",
				"tags": [
					"Dashboard"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "startDate",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "endDate",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "timeRange",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"1w",
								"1m",
								"3m",
								"6m",
								"1y"
							]
						}
					}
				]
			}
		},
		"/dashboard/{userId}/goal-progress": {
			"get": {
				"operationId": "GetGoalProgress",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GoalProgressResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GoalProgressResponse"
								}
							}
						}
					}
				},
				"description": "Get goal progress for a user",
				"tags": [
					"Dashboard"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "startDate",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "endDate",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "timeRange",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"1w",
								"1m",
								"3m",
								"6m",
								"1y"
							]
						}
					}
				]
			}
		},
		"/dashboard/{userId}/total-volume": {
			"get": {
				"operationId": "GetTotalVolumeMetrics",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TotalVolumeMetricsResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TotalVolumeMetricsResponse"
								}
							}
						}
					}
				},
				"description": "Get total volume metrics for a user (strength progress over time)",
				"tags": [
					"Dashboard"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "startDate",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "endDate",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "timeRange",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"1w",
								"1m",
								"3m",
								"6m",
								"1y"
							]
						}
					}
				]
			}
		},
...
		"/dashboard/{userId}/weight-progression": {
			"get": {
				"operationId": "GetWeightProgressionMetrics",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"items": {
												"properties": {
													"label": {
														"type": "string"
													},
													"maxWeight": {
														"type": "number",
														"format": "double"
													},
													"avgWeight": {
														"type": "number",
														"format": "double"
													},
													"date": {
														"type": "string"
													}
												},
												"required": [
													"label",
													"maxWeight",
													"avgWeight",
													"date"
												],
												"type": "object"
											},
											"type": "array"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"items": {
												"properties": {
													"label": {
														"type": "string"
													},
													"maxWeight": {
														"type": "number",
														"format": "double"
													},
													"avgWeight": {
														"type": "number",
														"format": "double"
													},
													"date": {
														"type": "string"
													}
												},
												"required": [
													"label",
													"maxWeight",
													"avgWeight",
													"date"
												],
												"type": "object"
											},
											"type": "array"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Get weight progression metrics for a user",
				"tags": [
					"Dashboard"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "startDate",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "endDate",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "timeRange",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"1w",
								"1m",
								"3m",
								"6m",
								"1y"
							]
						}
					}
				]
			}
		},
		"/dashboard/{userId}/weight-accuracy-by-date": {
			"get": {
				"operationId": "GetWeightAccuracyByDate",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"items": {
												"properties": {
													"label": {
														"type": "string"
													},
													"lowerWeight": {
														"type": "number",
														"format": "double"
													},
													"higherWeight": {
														"type": "number",
														"format": "double"
													},
													"exactMatches": {
														"type": "number",
														"format": "double"
													},
													"totalSets": {
														"type": "number",
														"format": "double"
													},
													"date": {
														"type": "string"
													}
												},
												"required": [
													"label",
													"lowerWeight",
													"higherWeight",
													"exactMatches",
													"totalSets",
													"date"
												],
												"type": "object"
											},
											"type": "array"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"items": {
												"properties": {
													"label": {
														"type": "string"
													},
													"lowerWeight": {
														"type": "number",
														"format": "double"
													},
													"higherWeight": {
														"type": "number",
														"format": "double"
													},
													"exactMatches": {
														"type": "number",
														"format": "double"
													},
													"totalSets": {
														"type": "number",
														"format": "double"
													},
													"date": {
														"type": "string"
													}
												},
												"required": [
													"label",
													"lowerWeight",
													"higherWeight",
													"exactMatches",
													"totalSets",
													"date"
												],
												"type": "object"
											},
											"type": "array"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Get raw weight accuracy data by date for frontend filtering",
				"tags": [
					"Dashboard"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "startDate",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "endDate",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "timeRange",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"1w",
								"1m",
								"3m",
								"6m",
								"1y"
							]
						}
					}
				]
			}
		},
		"/dashboard/{userId}/workout-type-by-date": {
			"get": {
				"operationId": "GetWorkoutTypeByDate",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"items": {
												"properties": {
													"label": {
														"type": "string"
													},
													"workoutTypes": {
														"items": {
															"properties": {
																"exerciseCount": {
																	"type": "number",
																	"format": "double"
																},
																"totalReps": {
																	"type": "number",
																	"format": "double"
																},
																"totalSets": {
																	"type": "number",
																	"format": "double"
																},
																"label": {
																	"type": "string"
																},
																"tag": {
																	"type": "string"
																}
															},
															"required": [
																"exerciseCount",
																"totalReps",
																"totalSets",
																"label",
																"tag"
															],
															"type": "object"
														},
														"type": "array"
													},
													"date": {
														"type": "string"
													}
												},
												"required": [
													"label",
													"workoutTypes",
													"date"
												],
												"type": "object"
											},
											"type": "array"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"items": {
												"properties": {
													"label": {
														"type": "string"
													},
													"workoutTypes": {
														"items": {
															"properties": {
																"exerciseCount": {
																	"type": "number",
																	"format": "double"
																},
																"totalReps": {
																	"type": "number",
																	"format": "double"
																},
																"totalSets": {
																	"type": "number",
																	"format": "double"
																},
																"label": {
																	"type": "string"
																},
																"tag": {
																	"type": "string"
																}
															},
															"required": [
																"exerciseCount",
																"totalReps",
																"totalSets",
																"label",
																"tag"
															],
															"type": "object"
														},
														"type": "array"
													},
													"date": {
														"type": "string"
													}
												},
												"required": [
													"label",
													"workoutTypes",
													"date"
												],
												"type": "object"
											},
											"type": "array"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Get raw workout type data by date for frontend filtering",
				"tags": [
					"Dashboard"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "startDate",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "endDate",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "timeRange",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"1w",
								"1m",
								"3m",
								"6m",
								"1y"
							]
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}